fix duplicate RestMapperTest issue #8
